Implements Custom JSON Root Feature
This feature implements behavior for allowing a user to specify a
custom root path for all JSON settings for each project. This works
by allowing the user to specify the Root path, and the Solution name
and Project names are added to the path.

This allows for saving modified Command Line arguments in a Source
Control solution, while having an ephemeral Visual Studio project
(such as the case when the VS projects are generated by a build
system). Due to this, this resolves issue #115

* Adds settings for "Use Custom JSON Root" and "Custom JSON
  Root"
* These options are hidden unless the "Save JSON Settings"
  setting is true.